Fix: Wrap useSearchParams in Suspense boundary for /studio page
The `useSearchParams()` hook in your `/studio` page was causing a build error
("⨯ useSearchParams() should be wrapped in a suspense boundary") because
it was used in a client-side rendered component without a Suspense boundary.

This commit resolves the issue by:
1. Extracting the original content of `app/studio/page.tsx` into a new
   component, `app/studio/StudioPageContent.tsx`.
2. Modifying `app/studio/page.tsx` to render `StudioPageContent`
   wrapped in a `<Suspense>` component with a fallback UI.

This ensures that the component utilizing `useSearchParams` is correctly
managed by React Suspense, preventing the build error.
